requirements.txt
###### External dependencies ######
PyYAML
jsonschema
prettytable
pyparsing
sqlalchemy

###### Test dependencies ######
pytest
pytest-cov
pylint